+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - OpenapiFirst::Test.report_coverage now includes fractional digits when returning a coverage value to avoid reporting "0% / no requests made" even though some requests have been made.
 - Add option to skip certain responses in coverage calculation
+- Show details about invalid requests / responses in coverage report
 
 ## 2.4.0